proper 1.21.9 port utilizing vanilla api more

+++ common/src/main/java/io/github/fishstiz/cursors_extended/config/AbstractCursorSettings.java
@@ -3,14 +3,14 @@ package io.github.fishstiz.cursors_extended.config;
 import io.github.fishstiz.cursors_extended.util.SettingsUtil;
 
 public abstract class AbstractCursorSettings<T extends AbstractCursorSettings<T>> {
-    protected double scale = SettingsUtil.SCALE;
+    protected float scale = SettingsUtil.SCALE;
     protected int xhot = SettingsUtil.X_HOT;
     protected int yhot = SettingsUtil.Y_HOT;
 
     protected AbstractCursorSettings() {
     }
 
-    public double getScale() {
+    public float getScale() {
         return scale;
     }
